+++ modal-js/test/sandbox.test.ts
@@ -4,6 +4,7 @@ import { buildSandboxCreateRequestProto } from "../src/sandbox";
 import { expect, test, onTestFinished } from "vitest";
 import { buildContainerExecRequestProto } from "../src/sandbox";
 import { GPUConfig, PTYInfo_PTYType } from "../proto/modal_proto/api";
+import { g } from "vitest/dist/chunks/suite.d.FvehnV49.js";
 
 test("CreateOneSandbox", async () => {
   const app = await tc.apps.fromName("libmodal-test", {
@@ -623,3 +624,19 @@ test("buildSandboxCreateRequestProto negative Memory", async () => {
     }),
   ).rejects.toThrow("must be a positive number");
 });
+
+test("ConnectToken", async () => {
+  const app = await tc.apps.fromName("libmodal-test", {
+    createIfMissing: true,
+  });
+  const image = tc.images.fromRegistry("python:3.12-alpine");
+
+  const sb = await tc.sandboxes.create(app, image);
+  onTestFinished(async () => {
+    await sb.terminate();
+  });
+
+  const creds = await sb.createConnectToken({userMetadata: "abc"});
+  expect(creds.token).toBeTruthy();
+  expect(creds.url).toBeTruthy();
+})